+++ jvm/controls/src/main/kotlin/org/wycliffeassociates/otter/jvm/controls/card/WizardCard.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.jvm.controls
+package org.wycliffeassociates.otter.jvm.controls.card
 
 import com.jfoenix.controls.JFXButton
 import javafx.beans.property.SimpleObjectProperty
