+++ jvm/controls/src/main/kotlin/org/wycliffeassociates/otter/jvm/controls/navigation/ProjectNav.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.jvm.controls.projectnav
+package org.wycliffeassociates.otter.jvm.controls.navigation
 
 import javafx.beans.property.SimpleStringProperty
 import javafx.geometry.Pos
