+++ jvm/controls/src/main/kotlin/org/wycliffeassociates/otter/jvm/controls/skins/media/AudioPlayerSkin.kt
@@ -8,7 +8,7 @@ import javafx.scene.control.SkinBase
 import javafx.scene.control.Slider
 import javafx.scene.layout.HBox
 import org.kordamp.ikonli.javafx.FontIcon
-import org.wycliffeassociates.otter.jvm.controls.AudioPlayerNode
+import org.wycliffeassociates.otter.jvm.controls.media.AudioPlayerNode
 import org.wycliffeassociates.otter.jvm.controls.controllers.AudioPlayerController
 import org.wycliffeassociates.otter.jvm.utils.onChangeAndDoNow
 import tornadofx.*
