+++ jvm/controls/src/main/kotlin/org/wycliffeassociates/otter/jvm/controls/styles/AppStyles.kt
@@ -7,7 +7,6 @@ import javafx.scene.paint.Color
 import javafx.scene.text.Font
 import javafx.scene.text.FontWeight
 import org.wycliffeassociates.otter.jvm.utils.images.ImageLoader
-import org.wycliffeassociates.otter.jvm.controls.progressdialog.ProgressDialogStyles
 import tornadofx.*
 
 class AppStyles : Stylesheet() {
