+++ jvm/controls/src/main/kotlin/org/wycliffeassociates/otter/jvm/controls/styles/ProgressDialogStyles.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.jvm.controls.progressdialog
+package org.wycliffeassociates.otter.jvm.controls.styles
 
 import javafx.geometry.Pos
 import javafx.scene.text.FontWeight
