+++ jvm/controls/src/main/kotlin/org/wycliffeassociates/otter/jvm/controls/toggle/SourceFormatToggle.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.jvm.controls.sourceformattoggle
+package org.wycliffeassociates.otter.jvm.controls.toggle
 
 import javafx.beans.property.SimpleObjectProperty
 import javafx.scene.control.Control
