+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/theme/AppStyles.kt
@@ -7,7 +7,7 @@ import javafx.scene.paint.Color
 import javafx.scene.text.Font
 import javafx.scene.text.FontWeight
 import org.wycliffeassociates.otter.jvm.utils.images.ImageLoader
-import org.wycliffeassociates.otter.jvm.controls.progressdialog.ProgressDialogStyles
+import org.wycliffeassociates.otter.jvm.controls.styles.ProgressDialogStyles
 import tornadofx.*
 
 class AppStyles : Stylesheet() {
