+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/projectwizard/view/fragments/SelectLanguage.kt
@@ -3,10 +3,10 @@ package org.wycliffeassociates.otter.jvm.workbookapp.ui.projectwizard.view.fragm
 import javafx.beans.property.Property
 import org.slf4j.LoggerFactory
 import org.wycliffeassociates.otter.common.data.primitives.Language
-import org.wycliffeassociates.otter.jvm.controls.searchablelist.SearchableList
+import org.wycliffeassociates.otter.jvm.controls.listview.SearchableList
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.projectwizard.view.ProjectWizardStyles
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.projectwizard.viewmodel.ProjectWizardViewModel
-import org.wycliffeassociates.otter.jvm.controls.searchablelist.searchablelist
+import org.wycliffeassociates.otter.jvm.controls.listview.searchablelist
 import tornadofx.*
 
 class SelectLanguage : Fragment() {
